+++ src/withLocalStore.ts
@@ -2,6 +2,6 @@ import * as React from 'react'
 import createConnect from './createConnect'
 import localStorageStore from './localStorageStore'
 
-export default function withLocalStorage (mapStoreToValues) {
-  return createConnect(mapStoreToValues, localStorageStore)
+export default function withLocalStorage (mapStoreToValues, handlers) {
+  return createConnect(mapStoreToValues, handlers, localStorageStore)
 }