+++ src/withSessionStore.ts
@@ -2,6 +2,6 @@ import * as React from 'react'
 import createConnect from './createConnect'
 import sessionStorageStore from './sessionStorageStore'
 
-export default function withSessionStorage (mapStoreToValues) {
-  return createConnect(mapStoreToValues, sessionStorageStore)
+export default function withSessionStorage (mapStoreToValues, handlers) {
+  return createConnect(mapStoreToValues, handlers, sessionStorageStore)
 }